[bsp][essemi] Adapt ES32F369x for RTduino
