+++ bsp/essemi/es32f369x/applications/main.c
@@ -17,7 +17,6 @@
 #else
 #define LED_PIN    GET_PIN( F , 0 )
 #endif
-          
 
 int main(void)
 {
@@ -29,9 +28,8 @@ int main(void)
     {
         rt_pin_write(LED_PIN, PIN_HIGH);
         rt_thread_mdelay(500);
-        rt_pin_write(LED_PIN, PIN_LOW); 
+        rt_pin_write(LED_PIN, PIN_LOW);
         rt_thread_mdelay(500);
     }
     return RT_EOK;
 }
-
